requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
djangorestframework==3.13.1
drf-extra-fields==3.4.0
Pillow==9.1.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
